rand_distr: add module beta
